requirements.txt
sqlalchemy
alembic
psycopg2-binary
flask
werkzeug
flask-sqlalchemy
flask-cors
pillow
